backend/requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
dj-database-url
django-cors-headers